+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.4
+
+* Add `katja_vmstats:get_timer/1`, `katja_vmstats:start_timer/2` and `katja_vmstats:stop_timer/1` to programmatically interact with timers/intervals
+
 ## 0.3
 
 [Documentation](http://katja_vmstats.nifoc.pw/0.3/)